requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
pyyaml==6.0
urllib3==1.26.8
